refactor: move line_column_at onto MietteSpanContents
The line/column derivation added for the read_span optimization was a free
function in the graphical renderer that took four accessors straight off the
`SpanContents` payload. Make it a `pub(crate)` method on `MietteSpanContents`,
co-located in `source_impls.rs` with `context_info` — the newline handling and
`memchr2` scan it mirrors — so the call site collapses to
`contents.line_column_at(offset)` and the CRLF / out-of-bounds invariants live
next to their source of truth.

No behavior change; the differential test moves along and now runs on a plain
`cargo test` (it only needs `read_span`, not the graphical feature).
