+++ src/source_impls.rs
@@ -5,7 +5,7 @@ Default trait implementations for [`SourceCode`].
 use std::str::from_utf8;
 use std::{borrow::Cow, collections::VecDeque, fmt::Debug, sync::Arc};
 
-use crate::{MietteError, MietteSpanContents, SourceCode, SourceSpan};
+use crate::{MietteError, MietteSpanContents, SourceCode, SourceSpan, SpanContents};
 
 fn context_info<'a>(
     input: &'a [u8],
@@ -121,6 +121,66 @@ fn context_info<'a>(
         ))
     } else {
         Err(MietteError::OutOfBounds)
+    }
+}
+
+impl MietteSpanContents<'_> {
+    /// The 0-indexed line and column of an absolute source `offset` that lies
+    /// within this payload, derived without re-reading the source.
+    ///
+    /// Equivalent to the `line()`/`column()` a fresh
+    /// `read_span(&(offset, 0).into(), 0, 0)` would report, but obtained by
+    /// scanning only this payload's prefix up to `offset` — letting a renderer
+    /// that already holds a `SpanContents` locate a label inside it instead of
+    /// issuing a second full [`SourceCode::read_span`]. Returns `None` when
+    /// `offset` is past this payload (which `read_span` reports as
+    /// `OutOfBounds`). Newline handling mirrors [`context_info`] (a `\r\n` pair
+    /// and a lone `\r`/`\n` each count once), scanned with the same `memchr2`
+    /// primitive so a long (e.g. minified) line stays cheap.
+    // Only the `fancy` graphical renderer consumes this today; without it the
+    // method is dead in a lib-only build (CI lints `-D warnings`).
+    #[cfg_attr(not(feature = "fancy-base"), allow(dead_code))]
+    pub(crate) fn line_column_at(&self, offset: usize) -> Option<(usize, usize)> {
+        let data = self.data();
+        let base = self.span().offset() as usize;
+        let mut rel = offset.saturating_sub(base);
+        // A label past the end of the payload is out of bounds; report it like
+        // the `OutOfBounds` the equivalent `read_span` would return, rather than
+        // clamping to a misleading end-of-snippet position.
+        if rel > data.len() {
+            return None;
+        }
+        // An offset landing on the `\n` of a `\r\n` pair sits inside an
+        // unfinished break: `context_info` has not advanced the line at that
+        // byte and reports the preceding `\r`. Normalize to that byte.
+        if rel > 0 && rel < data.len() && data[rel - 1] == b'\r' && data[rel] == b'\n' {
+            rel -= 1;
+        }
+        let mut line = self.line();
+        // Byte index just past the most recent line break — the start of the
+        // line `offset` falls on. `None` until the first break is seen, meaning
+        // `offset` is still on this payload's first line.
+        let mut line_start: Option<usize> = None;
+        for pos in memchr::memchr2_iter(b'\r', b'\n', &data[..rel]) {
+            // Skip the `\n` of a `\r\n` pair already counted at its `\r`.
+            if data[pos] == b'\n' && pos > 0 && data[pos - 1] == b'\r' {
+                continue;
+            }
+            line += 1;
+            // A `\r\n` pair is a single break ending at the `\n`.
+            let line_end = if data[pos] == b'\r' && pos + 1 < rel && data[pos + 1] == b'\n' {
+                pos + 1
+            } else {
+                pos
+            };
+            line_start = Some(line_end + 1);
+        }
+        Some(match line_start {
+            // A later line, which by definition starts at column 0.
+            Some(start) => (line, rel - start),
+            // Still on the first line: offset from this payload's start column.
+            None => (line, self.column() + rel),
+        })
     }
 }
 
@@ -332,5 +392,93 @@ mod tests {
         let span: SourceSpan = (0, 9).into();
         assert_eq!(&span, contents.span());
         Ok(())
+    }
+}
+
+#[cfg(test)]
+mod line_column_tests {
+    use super::*;
+
+    /// Deterministic xorshift so any failure reproduces from a fixed seed.
+    struct Rng(u64);
+    impl Rng {
+        fn next(&mut self) -> u64 {
+            let mut x = self.0;
+            x ^= x << 13;
+            x ^= x >> 7;
+            x ^= x << 17;
+            self.0 = x;
+            x
+        }
+        fn below(&mut self, n: usize) -> usize {
+            (self.next() % n as u64) as usize
+        }
+    }
+
+    /// Assert `MietteSpanContents::line_column_at` matches a real
+    /// `read_span(&(off, 0), 0, 0)` for an offset inside a
+    /// `read_span(&(off, len), ctx, ctx)` payload — exactly how the graphical
+    /// renderer consumes it. Full equivalence, *including* the error case:
+    /// `line_column_at` must return `None` iff the direct read is `OutOfBounds`
+    /// (a label past the snippet, e.g. `off == src.len() + 1`). Returns whether
+    /// a case was actually checked (the context read succeeded).
+    fn check(src: &str, off: usize, len: usize, ctx: usize) -> bool {
+        let Ok(contents) = src.read_span(&(off as u32, len as u32).into(), ctx, ctx) else {
+            return false;
+        };
+        let got = contents.line_column_at(off);
+
+        match src.read_span(&(off as u32, 0u32).into(), 0, 0) {
+            Ok(expected) => assert_eq!(
+                got,
+                Some((expected.line(), expected.column())),
+                "mismatch for src={src:?} off={off} len={len} ctx={ctx}"
+            ),
+            Err(_) => assert_eq!(
+                got, None,
+                "accepted an out-of-bounds label for src={src:?} off={off} len={len} ctx={ctx}"
+            ),
+        }
+        true
+    }
+
+    /// Exhaustively check every char-boundary offset of many small randomized
+    /// sources — plus offsets one and two bytes past EOF — across LF / CRLF /
+    /// lone-CR and multibyte alphabets.
+    #[test]
+    fn matches_read_span_exhaustively() {
+        let alphabets: &[&[&str]] = &[
+            &["a", "\n"],
+            &["a", "b", "c", "\n"],
+            &["a", "b", "\r\n"],
+            &["a", "\r", "\n", "\r\n"],
+            &["x", "y", "\n", "é", "🦀", "\r\n"],
+        ];
+        let mut rng = Rng(0x9E37_79B9_7F4A_7C15);
+        let mut checked = 0usize;
+        for alpha in alphabets {
+            for _ in 0..3000 {
+                let n = rng.below(14);
+                let mut s = String::new();
+                for _ in 0..n {
+                    s.push_str(alpha[rng.below(alpha.len())]);
+                }
+                // Include `len + 1` / `len + 2`, which are past EOF (never char
+                // boundaries) to exercise the out-of-bounds rejection.
+                for off in 0..=s.len() + 2 {
+                    if off <= s.len() && !s.is_char_boundary(off) {
+                        continue;
+                    }
+                    for ctx in 0..=2 {
+                        for &len in &[0usize, 1, 4] {
+                            if check(&s, off, len, ctx) {
+                                checked += 1;
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        assert!(checked > 100_000, "expected a broad sweep, only checked {checked}");
     }
 }
